+++ Source/Starbreeze/Public/SBZEnvQueryGenerator_WorldRuntimeActors.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "CoreMinimal.h"
-//CROSS-MODULE INCLUDE V2: -ModuleName=AIModule -ObjectName=EnvQueryGenerator -FallbackName=EnvQueryGenerator
+#include "EnvironmentQuery/EnvQueryGenerator.h"
 #include "SBZEnvQueryGenerator_WorldRuntimeActors.generated.h"
 
 UCLASS(Blueprintable, EditInlineNew)
@@ -9,9 +9,10 @@ class USBZEnvQueryGenerator_WorldRuntimeActors : public UEnvQueryGenerator {
 public:
 protected:
     UPROPERTY(EditAnywhere, meta=(AllowPrivateAccess=true))
-    uint16 ActorSetting;
+    uint32 ActorSetting;
     
 public:
     USBZEnvQueryGenerator_WorldRuntimeActors();
+
 };
